Remove unneeded ResponseAwareTrait
